start reorganizing FFI
