+++ wasm/environment.go
@@ -2,31 +2,10 @@ package wasm
 
 import (
 	"sync"
-	"unsafe"
 
 	"github.com/pkg/errors"
 	"github.com/wasmerio/wasmer-go/wasmer"
 )
-
-// #include <stdlib.h>
-//
-// extern void return_result(void *context, int32_t pointer, int32_t size, int32_t envIndex, int32_t instIndex);
-import "C"
-
-/*
- In order to allow "easy" communication of data across the FFI barrier (outbound Go -> WASM and inbound WASM -> Go), hivew provides
- an FFI API. Functions exported from a WASM module can be easily called by Go code via the Wasmer instance exports, but returning data
- to the host Go code is not quite as straightforward.
-
- In order to accomplish this, hivew internally keeps a set of "environments" in a singleton package var (`environments` below).
- Each environment is a container that includes the WASM module bytes, and a set of WASM instances (runtimes) to execute said module.
- The envionment object has an index referencing its place in the singleton array, and each instance has an index referencing its position within
- the environment's instance array.
-
- When a WASM function calls one of the FFI API functions, it includes the `env_index` and `inst_index` values that were provided at the beginning
- of job execution, which allows hivew to look up the [env][instance] and send the result on the appropriate result channel. This is needed due to
- the way Go makes functions available on the FFI using CGO.
-*/
 
 var environments []*wasmEnvironment
 var envLock = sync.RWMutex{}
@@ -124,55 +103,23 @@ func (w *wasmEnvironment) addInstance() error {
 	return nil
 }
 
+func instanceAtIndices(envIndex int32, instIndex int32) *wasmInstance {
+	if int(envIndex) > len(environments)-1 {
+		return nil
+	}
+
+	env := environments[envIndex]
+
+	if int(instIndex) > len(env.instances)-1 {
+		return nil
+	}
+
+	return env.instances[instIndex]
+}
+
 // setRaw sets the raw bytes of a WASM module to be used rather than a filepath
 func (w *wasmEnvironment) setRaw(raw []byte) {
 	w.raw = raw
-}
-
-func (w *wasmInstance) writeInput(input []byte) int32 {
-	lengthOfInput := len(input)
-
-	// Allocate memory for the input, and get a pointer to it.
-	allocateResult, _ := w.wasmerInst.Exports["allocate_input"](lengthOfInput)
-	inputPointer := allocateResult.ToI32()
-
-	// Write the input into the memory.
-	memory := w.wasmerInst.Memory.Data()[inputPointer:]
-
-	for index := 0; index < lengthOfInput; index++ {
-		memory[index] = input[index]
-	}
-
-	return inputPointer
-}
-
-func (w *wasmInstance) deallocate(pointer int32, length int) {
-	dealloc := w.wasmerInst.Exports["deallocate"]
-
-	dealloc(pointer, length)
-}
-
-///////////////////////////////////////////////////////////////
-// below is the "hivew API" used for cross-FFI communication //
-///////////////////////////////////////////////////////////////
-
-//export return_result
-func return_result(context unsafe.Pointer, pointer int32, size int32, envIndex int32, instIndex int32) {
-	envLock.RLock()
-	defer envLock.RUnlock()
-
-	env := environments[envIndex]
-	inst := env.instances[instIndex]
-
-	memory := inst.wasmerInst.Memory.Data()[pointer:]
-
-	result := make([]byte, size)
-
-	for index := 0; int32(index) < size; index++ {
-		result[index] = memory[index]
-	}
-
-	inst.resultChan <- result
 }
 
 func init() {
